requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
tensorflow==2.18.0
sqlmodel==0.0.21
alembic==1.13.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
pydantic-settings==2.6.0
pillow==11.0.0
numpy==1.26.4
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
jinja2==3.1.4
aiofiles==24.1.0
python-dotenv==1.0.0
